+++ src/sdk/PnP.Core/Model/SharePoint/Taxonomy/Internal/TermStore.cs
@@ -1,4 +1,9 @@
-﻿using System.Collections.Generic;
+﻿using PnP.Core.Services;
+using System;
+using System.Collections.Generic;
+using System.Linq.Expressions;
+using System.Net.Http;
+using System.Threading.Tasks;
 
 namespace PnP.Core.Model.SharePoint
 {
@@ -19,6 +24,68 @@ namespace PnP.Core.Model.SharePoint
 
         [KeyProperty(nameof(Id))]
         public override object Key { get => Id; set => Id = value.ToString(); }
+        #endregion
+
+        #region Methods
+
+        public async Task<ITermSet> GetTermSetByIdAsync(string id, params Expression<Func<ITermSet, object>>[] selectors)
+        {
+            var termSet = new TermSet()
+            {
+                PnPContext = PnPContext
+            };
+
+            var apiCall = BuildGetTermSetByIdApiCall(id);
+
+            var entityInfo = EntityManager.GetClassInfo(termSet.GetType(), termSet, expressions: selectors);
+            var query = await QueryClient.BuildGetAPICallAsync(termSet, entityInfo, apiCall).ConfigureAwait(false);
+
+            await termSet.RequestAsync(new ApiCall(query.ApiCall.Request, ApiType.Graph), HttpMethod.Get).ConfigureAwait(false);
+
+            return termSet;
+        }
+
+        public ITermSet GetTermSetById(string id, params Expression<Func<ITermSet, object>>[] selectors)
+        {
+            return GetTermSetByIdAsync(id, selectors).GetAwaiter().GetResult();
+        }
+
+        public async Task<ITermSet> GetTermSetByIdBatchAsync(string id, params Expression<Func<ITermSet, object>>[] selectors)
+        {
+            return await GetTermSetByIdBatchAsync(PnPContext.CurrentBatch, id, selectors).ConfigureAwait(false);
+        }
+
+        public ITermSet GetTermSetByIdBatch(string id, params Expression<Func<ITermSet, object>>[] selectors)
+        {
+            return GetTermSetByIdBatchAsync(id, selectors).GetAwaiter().GetResult();
+        }
+
+        public async Task<ITermSet> GetTermSetByIdBatchAsync(Batch batch, string id, params Expression<Func<ITermSet, object>>[] selectors)
+        {
+            var termSet = new TermSet()
+            {
+                PnPContext = PnPContext
+            };
+
+            var apiCall = BuildGetTermSetByIdApiCall(id);
+
+            var entityInfo = EntityManager.GetClassInfo(termSet.GetType(), termSet, expressions: selectors);
+            var query = await QueryClient.BuildGetAPICallAsync(termSet, entityInfo, apiCall).ConfigureAwait(false);
+
+            await termSet.RequestBatchAsync(batch, new ApiCall(query.ApiCall.Request, ApiType.Graph), HttpMethod.Get).ConfigureAwait(false);
+
+            return termSet;
+        }
+
+        public ITermSet GetTermSetByIdBatch(Batch batch, string id, params Expression<Func<ITermSet, object>>[] selectors)
+        {
+            return GetTermSetByIdBatchAsync(batch, id, selectors).GetAwaiter().GetResult();
+        }
+
+        private ApiCall BuildGetTermSetByIdApiCall(string termSetId)
+        {
+            return new ApiCall($"sites/{PnPContext.Site.Id}/termstore/sets/{termSetId}", ApiType.Graph);
+        }
         #endregion
     }
 }